Reorganize the stories
