+++ packages/veritone-react-common/src/components/DataSets/DataSetFullscreenViewer/styles.scss
@@ -23,27 +23,29 @@
     .helpIcon {
       height: 20px;
       width: 20px;
-      color: #FAFAFA;
+      font-size: 20px;
+      color: #e0e0e0;
       align-self: center;
     }
     .menuIcon {
-      height: 20px;
-      width: 20px;
-      color: #FAFAFA;
+      height: 22px;
+      width: 22px;
+      font-size: 22px;
+      color: #e0e0e0;
       align-self: center;
     }
     .separator {
       align-self: center;
-      color: #FAFAFA;
-      border-right: 2px solid white;
+      border-right: 2px solid #e0e0e0;
       height: 27px;
       margin-left: -2px;
       margin-bottom: -2px;
     }
     .exitIcon {
-      height: 20px;
-      width: 20px;
-      color: #FAFAFA;
+      height: 22px;
+      width: 22px;
+      font-size: 22px;
+      color: #e0e0e0;
       align-self: center;
     }
   }
