+++ packages/veritone-react-common/src/components/SDO/SDOCard/styles.scss
@@ -39,7 +39,7 @@
     align-self: center;
   }
   .menuIcon {
-    font-size: 16px;
+    font-size: 20px;
     align-self: center;
   }
 }
